Remove old API and implementation
