--- lib/rummageable/fake.rb
@@ -1,18 +1,0 @@
-module Rummageable
-  class Fake
-    def index(documents, index_name)
-    end
-
-    def delete(link, index_name)
-    end
-
-    def amend(link, amendments, index_name)
-    end
-
-    def commit(index_name)
-    end
-
-    def validate_structure(hash, parents=[])
-    end
-  end
-end
